modify stock and warehouse

+++ CarWashManagementSystem/Warehouse.xaml.cs
@@ -14,18 +14,19 @@ namespace CarWashManagementSystem
     {
         ProductService _productService;
         ValidationService _validation;
+        bool _isDiscontinued = false;
         public Warehouse()
         {
             InitializeComponent();
             _productService = new ProductService();
             _validation = new ValidationService();
-            ShowData();
+            ShowData(_isDiscontinued);
         }
 
-        private void ShowData()
+        private void ShowData(bool isDiscontinued)
         {
             dgvProduct.ItemsSource = null;
-            dgvProduct.ItemsSource = _productService.GetProducts();
+            dgvProduct.ItemsSource = _productService.GetProductsByStatus(isDiscontinued);
         }
 
         private void txtSearch_TextChanged(object sender, TextChangedEventArgs e)
@@ -58,7 +59,7 @@ namespace CarWashManagementSystem
             //    return;
             //}
 
-            if (!_validation.IsNumber(product.Price+""))
+            if (!_validation.IsNumber(product.Price + ""))
             {
                 MessageBox.Show("Product price must be a number.");
                 e.Cancel = true;
@@ -87,7 +88,7 @@ namespace CarWashManagementSystem
             }
 
             _productService.UpdateProduct(product);
-            ShowData();
+            ShowData(_isDiscontinued);
         }
 
         private void Add_Click(object sender, RoutedEventArgs e)
@@ -99,7 +100,7 @@ namespace CarWashManagementSystem
         }
         private void Module_Closed(object sender, EventArgs e)
         {
-            ShowData();
+            ShowData(_isDiscontinued);
         }
 
         private void Discontinued_Click(object sender, RoutedEventArgs e)
@@ -108,37 +109,86 @@ namespace CarWashManagementSystem
 
             Product product = (Product)btn.DataContext;
 
-            MessageBoxResult result = MessageBox.Show(
-                $"Are you sure to delete '{product.Name}'?",
+            if (product.IsDiscontinued)
+            {
+                MessageBoxResult result = MessageBox.Show(
+                $"Are you sure to move '{product.Name}' to actively product?",
                 "Confirm",
                 MessageBoxButton.YesNo,
                 MessageBoxImage.Warning
             );
 
-            if (result == MessageBoxResult.Yes)
-            {
-                try
+                if (result == MessageBoxResult.Yes)
                 {
-                    _productService.DeleteProduct(product);
+                    try
+                    {
+                        product.IsDiscontinued = false;
+                        _productService.UpdateProduct(product);
+                    }
+                    catch (Exception ex)
+                    {
+                        MessageBox.Show($"Error while processing: {ex.Message}");
+                    }
                 }
-                catch (Exception ex)
-                {
-                    MessageBox.Show($"Error while deleting product: {ex.Message}");
-                }
+                ShowData(_isDiscontinued);
             }
-            ShowData();
+            else
+            {
+
+                MessageBoxResult result = MessageBox.Show(
+                    $"Are you sure to move '{product.Name}' to discontinued product?",
+                    "Confirm",
+                    MessageBoxButton.YesNo,
+                    MessageBoxImage.Warning
+                );
+
+                if (result == MessageBoxResult.Yes)
+                {
+                    try
+                    {
+                        _productService.DeleteProduct(product);
+                    }
+                    catch (Exception ex)
+                    {
+                        MessageBox.Show($"Error while processing: {ex.Message}");
+                    }
+                }
+                ShowData(_isDiscontinued);
+            }
         }
 
         private void StockIn_Click(object sender, RoutedEventArgs e)
         {
             Button btn = (Button)sender;
 
-            Product product= (Product)btn.DataContext;
+            Product product = (Product)btn.DataContext;
+
+            if (product.IsDiscontinued)
+            {
+                MessageBox.Show("Reactive this product to do this.");
+                return;
+            }
 
             StockIn s = new StockIn(product);
-            s.WindowStartupLocation= WindowStartupLocation.CenterScreen;
+            s.WindowStartupLocation = WindowStartupLocation.CenterScreen;
             s.Closed += Module_Closed;
             s.ShowDialog();
+        }
+
+        private void ChangeView_Click(object sender, RoutedEventArgs e)
+        {
+            if (_isDiscontinued == false)
+            {
+                _isDiscontinued = true;
+                btnChangeView.Content = "Actively Product";
+            }
+            else
+            {
+                _isDiscontinued = false;
+                btnChangeView.Content = "Discontinued product";
+            }
+
+            ShowData(_isDiscontinued);
         }
     }
 }
